🌲 Trees

+++ src/main/java/com/james090500/blocks/Blocks.java
@@ -4,13 +4,13 @@ public class Blocks {
 
     public static Block[] ids;
 
-    public static final GrassBlock grassBlock = new GrassBlock(1);
-    public static final DirtBlock dirtBlock = new DirtBlock(2);
-    public static final StoneBlock stoneBlock = new StoneBlock(3);
-    public static final SandBlock sandBlock = new SandBlock(4);
-    public static final WaterBlock waterBlock = new WaterBlock(5);
-    public static final LogBlock logBlock = new LogBlock(6);
-    public static final LeafBlock leafBlock = new LeafBlock(7);
+    public static final GrassBlock grassBlock = new GrassBlock((byte) 1);
+    public static final DirtBlock dirtBlock = new DirtBlock((byte) 2);
+    public static final StoneBlock stoneBlock = new StoneBlock((byte) 3);
+    public static final SandBlock sandBlock = new SandBlock((byte) 4);
+    public static final WaterBlock waterBlock = new WaterBlock((byte) 5);
+    public static final LogBlock logBlock = new LogBlock((byte) 6);
+    public static final LeafBlock leafBlock = new LeafBlock((byte) 7);
 
     static {
         ids = new Block[] {
