+++ src/main/java/com/james090500/blocks/DirtBlock.java
@@ -1,7 +1,7 @@
 package com.james090500.blocks;
 
 public class DirtBlock extends Block {
-    public DirtBlock(int id) {
+    public DirtBlock(byte id) {
         super(id);
         this.texture = 2;
     }
