+++ src/main/java/com/james090500/blocks/LeafBlock.java
@@ -1,7 +1,7 @@
 package com.james090500.blocks;
 
 public class LeafBlock extends Block {
-    public LeafBlock(int id) {
+    public LeafBlock(byte id) {
         super(id);
         this.texture = 52;
         this.transparent = true;
