+++ src/main/java/com/james090500/blocks/SandBlock.java
@@ -1,7 +1,7 @@
 package com.james090500.blocks;
 
 public class SandBlock extends Block {
-    public SandBlock(int id) {
+    public SandBlock(byte id) {
         super(id);
         this.texture = 18;
     }
