+++ src/main/java/com/james090500/blocks/StoneBlock.java
@@ -2,7 +2,7 @@ package com.james090500.blocks;
 
 public class StoneBlock extends Block {
 
-    public StoneBlock(int id) {
+    public StoneBlock(byte id) {
         super(id);
         this.texture = 1;
     }
