+++ src/main/java/com/james090500/blocks/WaterBlock.java
@@ -1,7 +1,7 @@
 package com.james090500.blocks;
 
 public class WaterBlock extends Block {
-    public WaterBlock(int id) {
+    public WaterBlock(byte id) {
         super(id);
         this.texture = 205;
         this.transparent = true;
